xoroshiro32 conv fix

+++ generators/xoroshiro32pp.c
@@ -32,7 +32,7 @@ static inline uint16_t Xoroshiro32PPState_get_bits(Xoroshiro32PPState *obj)
 {
     const uint16_t s0 = obj->s[0];
     uint16_t s1 = obj->s[1];
-    const uint16_t result = rotl16((uint16_t)(s0 + s1), 9) + s0;
+    const uint16_t result = (uint16_t) (rotl16((uint16_t)(s0 + s1), 9) + s0);
     s1 ^= s0;
     obj->s[0] = (uint16_t) (rotl16(s0, 13) ^ s1 ^ (s1 << 5)); // a, b
     obj->s[1] = rotl16(s1, 10); // c
